// src/appointment/enums/messages.enum.ts
export enum Message {
    CREATED = 'Appointment created successfully',
    FIND_ONE = 'Appointment found successfully',
    FIND_ALL = 'Appointments found succesfully',
    UPDATED = 'Appointment updated succesfully',
    NOT_FOUND = `Appointment wasn't found`,
    ERROR = 'Something bad happened',
    TYPE_TRANSACTION = 'Appointment'
}